requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiogram==3.3.0
aiohttp==3.9.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
pyopenssl==23.3.0